chore: Update gin example to trigger long file paths
Added a second file to the gin example to demonstrate the problem with
file paths reported to in PLAT-203. With this configuration, the top
line of the stacktrace is:

/Users/[username]/path/to/go/code/src/github.com/bugsnag/bugsnag-go/examples/gin/request.go:31

We should instead trim content which can be derived by default, and
improve grouping when the Go application is built on different machines
or with different values for $GOPATH. The expected callstack frame for
the example error should then be:

request.go:31

Once $GOPATH is trimmed, adding an entry to ProjectPackages for the
current package should get the expected path to the file.
