+++ examples/gin/main.go
@@ -4,8 +4,6 @@ import (
 	"github.com/bugsnag/bugsnag-go"
 	"github.com/bugsnag/bugsnag-go/gin"
 	"github.com/gin-gonic/gin"
-	"net/http"
-	"os"
 )
 
 func main() {
@@ -14,29 +12,11 @@ func main() {
 
 	// Insert your API key
 	g.Use(bugsnaggin.AutoNotify(bugsnag.Configuration{
-		APIKey: "YOUR-API-KEY-HERE",
+		APIKey:          "YOUR-API-KEY-HERE",
+		ProjectPackages: []string{"main", "github.com/bugsnag/bugsnag-go/examples/gin"},
 	}))
 
-	g.GET("/crash", performUnhandledCrash)
-	g.GET("/handled", performHandledCrash)
+	ConfigureRoutes(g)
 
 	g.Run(":9001") // listen and serve on 0.0.0.0:9001
-}
-
-func performUnhandledCrash(c *gin.Context) {
-	c.String(http.StatusOK, "OK")
-	var a struct{}
-	crash(a)
-}
-
-func performHandledCrash(c *gin.Context) {
-	_, err := os.Open("some_nonexistent_file.txt")
-	if err != nil {
-		bugsnag.Notify(err)
-	}
-	c.String(http.StatusOK, "OK")
-}
-
-func crash(a interface{}) string {
-	return a.(string)
 }
